+++ src/pages/view/hooks/useKakaoMap.tsx
@@ -1,2 +1,142 @@
-export const useKakaoMap = () => {
-}
+import { useEffect, useState } from 'react';
+
+import {
+  ALLSTATIONLOCATIONS,
+  HONGDAELOCATIONS,
+  SAVEDLOCATIONS,
+} from 'src/constants/stations';
+
+import { useDebounce } from './useDebounce';
+
+import { CoordinateType, StationType } from '@types';
+
+const defaultCenter = { lat: 37.556621, lng: 126.923877 };
+
+export const useKakaoMap = (
+  currentLocation: StationType,
+  onMarkerClick: (storeId: number) => void
+) => {
+  const [storeMarkerList, setStoreMarkerList] = useState<
+    (CoordinateType & { clicked: boolean })[]
+  >([]);
+
+  // 현재 사용자의 위치
+  const [currentPosition, setCurrentPosition] = useState<{
+    lat: number;
+    lng: number;
+  }>(defaultCenter);
+
+  // 지도의 중심 좌표
+  const [center, setCenter] = useState<{
+    lat: number;
+    lng: number;
+  }>(defaultCenter);
+
+  // 찜 버튼 활성화 상태
+  const [isSaveActive, setIsSaveActive] = useState(false);
+
+  // gps 버튼 활성화 상태
+  const [isGpsActive, setIsGpsActive] = useState(false);
+
+  const fetchCurrentPosition = () => {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(
+        (position) => {
+          const { latitude, longitude } = position.coords;
+          const newCenter = { lat: latitude, lng: longitude };
+          setCurrentPosition(newCenter);
+          setCenter(newCenter);
+          setIsGpsActive(true);
+        },
+        (error) => {
+          console.error('Error getting user location:', error);
+          setCurrentPosition(defaultCenter);
+          setCenter(defaultCenter);
+          setIsGpsActive(false);
+        }
+      );
+    }
+  };
+
+  // useDebounce 훅 사용
+  const { debouncedFunction: handleDebouncedCenterChange } = useDebounce(
+    (map: kakao.maps.Map) => {
+      const lat = map.getCenter().getLat();
+      const lng = map.getCenter().getLng();
+      setCenter({ lat, lng });
+    },
+    500
+  );
+
+  const handleCenterChanged = (map: kakao.maps.Map) => {
+    setIsGpsActive(false);
+    handleDebouncedCenterChange(map);
+  };
+  const handleGpsButtonClick = () => {
+    fetchCurrentPosition();
+  };
+
+  const handleSaveButtonClick = () => {
+    setIsSaveActive((prev) => !prev);
+    setStoreMarkerList(
+      (isSaveActive ? ALLSTATIONLOCATIONS : SAVEDLOCATIONS).map((location) => ({
+        ...location,
+        clicked: false,
+      }))
+    );
+  };
+
+  const handleMarkerClick = (storeId: number) => {
+    setStoreMarkerList((prev) =>
+      prev.map((marker) =>
+        marker.storeId === storeId
+          ? { ...marker, clicked: true }
+          : { ...marker, clicked: false }
+      )
+    );
+    onMarkerClick(storeId);
+  };
+
+  const handleMapClick = () => {
+    // 지도 빈 공간 클릭 시 모든 마커의 clicked 상태를 초기화
+    setStoreMarkerList((prev) =>
+      prev.map((marker) => ({ ...marker, clicked: false }))
+    );
+  };
+
+  useEffect(() => {
+    if (currentLocation) {
+      if (currentLocation.stationEnName === 'ALL') {
+        fetchCurrentPosition();
+      } else {
+        setStoreMarkerList(
+          HONGDAELOCATIONS.map((location) => ({ ...location, clicked: false }))
+        );
+        setCenter({
+          lat: currentLocation.latitude,
+          lng: currentLocation.longitude,
+        });
+      }
+    }
+  }, [currentLocation]);
+
+  useEffect(() => {
+    fetchCurrentPosition();
+    setStoreMarkerList(
+      ALLSTATIONLOCATIONS.map((location) => ({ ...location, clicked: false }))
+    );
+  }, []);
+
+  return {
+    storeMarkerList,
+    currentPosition,
+    center,
+    isSaveActive,
+    isGpsActive,
+    handleCenterChanged,
+    handleGpsButtonClick,
+    handleSaveButtonClick,
+    handleMarkerClick,
+    handleMapClick,
+  };
+};
